# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kala API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/flashcards/{course_id}/deck": {
      "get": {
        "tags": [
          "flashcards"
        ],
        "summary": "Deck",
        "description": "The review queue: due scheduled cards first, topped up with fresh cards\nfor skills the student hasn't started. Answer keys never leave the server.\n`skill_id` is the topic picker's explicit override \u2014 when set, the whole\ndeck (due cards AND fresh top-up) scopes to that one skill instead of the\ncourse-wide due-first mix, same \"auto vs chosen\" pattern as practice.",
        "operationId": "deck_flashcards__course_id__deck_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "module_ref",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module Ref"
            }
          },
          {
            "name": "skill_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Skill Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/flashcards/{course_id}/review": {
      "post": {
        "tags": [
          "flashcards"
        ],
        "summary": "Review",
        "description": "Self-mark one study card and advance its spaced-repetition schedule.\n\nStudy mode is memorization, not assessment: the student flips the card,\nsees the answer, and marks \"Got it\" or \"Review again\" themselves. There is\nno server grading here \u2014 `remembered` is the student's own call, by design.\n\nWhat that means for the twin: this writes an evidence_event (type\n'flashcard', correct=remembered) but does NOT call tracer.apply_evidence,\nso mastery_state never moves on self-report. That row is still written on\npurpose \u2014 the research export (0003_research.sql) aggregates flashcard\nevidence into research_evidence_anon / research_cohort_daily, and dropping\nit would make every self-paced study session invisible to that pipeline.\nMastery is fed by diagnostic, quiz (test mode), and tutor checks only; the\nstudy -> test bridge is what routes studying into a real mastery signal.\nXP still moves off study because xp.summary reads evidence_events, which is\nthe intended behavior (reward the work, don't inflate mastery).",
        "operationId": "review_flashcards__course_id__review_post",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewBody": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "remembered": {
            "type": "boolean",
            "title": "Remembered"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "remembered"
        ],
        "title": "ReviewBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}